Fixed snow depth icon, restructured widget folders.
