+++ Site/bc_weather_station_dashboard/website/static/widgetsJS/fire/danger_rating.js
