+++ Site/bc_weather_station_dashboard/website/static/widgetsJS/weather/humidity.js
