+++ Site/bc_weather_station_dashboard/website/templates/fire.html
@@ -50,10 +50,10 @@
                 </div>
 
                 <!-- Leaflet Map Implementation -->
-                <div id="map" style="height: 53em; width: 100%;">
+                <div id="map" style="height: 52em; width: 100%;">
                     <button id="find-me-btn">Find me!</button>
                 </div>
-                <script src="{% static 'JavaScript/map.js' %}"></script>
+                <script src="{% static 'dataLogic/map.js' %}"></script>
                 <!-- Search bar -->
                 <div class="container mt-4">
                     <div class="input-group">
@@ -346,13 +346,13 @@
     </script>
 
     <!-- JS for widgets -->
-    <script src="{% static 'WidgetJavaScript/danger_rating.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/dmc.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/ffmc.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/dc.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/isi.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/bui.js' %}"></script>
-    <script src="{% static 'WidgetJavaScript/fwi.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/danger_rating.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/dmc.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/ffmc.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/dc.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/isi.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/bui.js' %}"></script>
+    <script src="{% static 'widgetsJS/fire/fwi.js' %}"></script>
 </body>
 
 </html>
