+++ Site/bc_weather_station_dashboard/website/templates/navbar.html
@@ -122,5 +122,5 @@
         </div>
     </div>
 </div>
-<script src="{% static 'JavaScript/map.js' %}"></script>
+<script src="{% static 'dataLogic/map.js' %}"></script>
 {% include "alerts.html" %}
