+++ Site/bc_weather_station_dashboard/website/templates/view_favourites.html
@@ -1,6 +1,6 @@
 {% load static %}
 <link rel="stylesheet" type="text/css" href="{% static 'CSS/favourite_cards.css' %}">
-<script src="../static/JavaScript/map.js"></script>
+<script src="../static/dataLogic/map.js"></script>
 
 
 <div class="modal" id="viewFavouritesModal">
