+++ Site/bc_weather_station_dashboard/website/tests/wind_directionTest.js
@@ -1,6 +1,6 @@
 const jsdom = require('jsdom');
 const { JSDOM } = jsdom;
-const WindArrow = require('../static/WidgetJavaScript/wind_direction.js');
+const WindArrow = require('../static/widgetsJS/weather/wind_direction.js');
 
 QUnit.module('WindArrow', function(hooks) {
     hooks.beforeEach(function() {
